Update README.md
To all my dear viewers, I've made some changes to the naming within this repository from 'OCP 4.10.3' to 'OpenShift 4.10'. This is to accommodate generic 4.10 installation on the GCP platform. This also means that when cloning this repository, the repo link will also be different. So please follow through the installation instructions carefully as they have been updated to accommodate the necessary changes. If anything is missing, feel free to let me know by creating an issue ticket and I will look into it.